+++ changelog.md
@@ -4,6 +4,7 @@ Liferay Plugin for IntelliJ
 All notable changes to this project will be documented in this file.
 
 ## 0.0.17
+- Feature: Code completion features for client-extensions.yaml
 - Change: Update support for IntelliJ 2023.1, dropped support for older versions
 
 ## 0.0.16